Extract termination persistence helpers from AppDelegate
