+++ Sources/TerminationSessionPersistencePlan.swift
@@ -1,0 +1,5 @@
+struct TerminationSessionPersistencePlan: Equatable, Sendable {
+    let saveSnapshot: Bool
+    let includeScrollback: Bool
+    let flushClosedItemHistory: Bool
+}
