+++ Sources/TerminationSessionPersistenceReason.swift
@@ -1,0 +1,6 @@
+enum TerminationSessionPersistenceReason: Sendable {
+    case applicationWillTerminate
+    case workspaceWillPowerOff
+    case sessionDidResignWhileTerminating
+    case updateRelaunch
+}
